jps 1차구현 완
TODO: path 통쨰로 뿌리거나, 일정갯수주고 몇번째 패스 도착했을 때 다음 패스 요청하게 수정 / 루트 몇개 되지도않는데 그냥 통쨰로 뿌리는게 좋을듯
TODO: path 이동할때 rotation lerp해서 수정해야 이쁠듯

+++ GameServer/GameServer/FieldPacketHandleThread.h
@@ -75,8 +75,8 @@ private:
 	//HandleCharacterMove 요청왔을때 길찾기하는 쓰레드 만들기
 	uint8** _map;
 	void InitializeMap();
-	uint8 _mapSizeX;
-	uint8 _mapSizeY;
+	uint32 _mapSizeX;
+	uint32 _mapSizeY;
 	JumpPointSearch* _jps;
